Hook up request and show for typing candidates in WK1 https://bugs.webkit.org/show_bug.cgi?id=151831 -and corresponding- <rdar://problem/23751214>
Reviewed by Enrica Casucci.

Source/WebCore:

New SPI that is needed.
* WebCore.xcodeproj/project.pbxproj:
* platform/spi/mac/NSSpellCheckerSPI.h: Added.

Request candidates for editable content whenever selection changes.
* editing/Editor.cpp:
(WebCore::Editor::respondToChangedSelection):

Implement requestCandidatesForSelection on the EditorClient.
* loader/EmptyClients.h:
* page/EditorClient.h:
(WebCore::EditorClient::requestCandidatesForSelection):

Source/WebKit/mac:

Add member variables to WebEditorClient. One is a WeakPtrFactory for the
asynchronous handlers, and the other caches the VisibleSelection at the time
candidates were requested so that we can make sure the candidates are still
valid once we receive them.
* WebCoreSupport/WebEditorClient.h:
* WebCoreSupport/WebEditorClient.mm:
(WebEditorClient::WebEditorClient):

Call [NSSpellChecker requestCandidatesForSelectedRange] with the appropriate
parameters.
(WebEditorClient::requestCandidatesForSelection):

Two helpers to compute information that we need for both of the handlers
below.
(candidateRangeForSelection):
(candidateWouldReplaceText):

In this handler, we just need to call [NSSpellChecker showCandidates] with
the appropriate parameters.
(WebEditorClient::handleRequestedCandidates):

Once a candidate is accepted, it should be inserted in the right way.
(WebEditorClient::handleAcceptedCandidate):

Source/WebKit2:

Empty client for now.
* WebProcess/WebCoreSupport/WebEditorClient.h:



Identifier: 170095@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@193645 268f45cc-cd09-0410-ab3c-d52691b4dbfc