Web Inspector: Remove "local" scope type from the protocol https://bugs.webkit.org/show_bug.cgi?id=152409
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-12-17
Reviewed by Timothy Hatcher.

Source/JavaScriptCore:

After r194251 the backend no longer sends this scope type.
So remove it from the protocol.

The concept of a Local Scope should be calculatable by the
frontend. In fact the way the backend used to do this could
easily be done by the frontend. To be done in a follow-up.

* inspector/InjectedScriptSource.js:
* inspector/JSJavaScriptCallFrame.h:
* inspector/protocol/Debugger.json:

Source/WebInspectorUI:

* UserInterface/Controllers/DebuggerManager.js:
(WebInspector.DebuggerManager.prototype._scopeChainNodeFromPayload):

Identifier: 170526@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@194259 268f45cc-cd09-0410-ab3c-d52691b4dbfc