Float with media query positioned incorrectly after window resize. https://bugs.webkit.org/show_bug.cgi?id=152558
Reviewed by Simon Fraser and David Hyatt.

This patch ensures that when a renderer becomes floated, it is moved
to the right containing block.
When this floated renderer's previous sibling is an anonymous block, it needs
to be reparented so that the float is positioned as if there was no anonymous block at all.

Source/WebCore:

Test: fast/block/float/float-with-anonymous-previous-sibling.html

* rendering/RenderElement.cpp:
(WebCore::RenderElement::styleDidChange):

LayoutTests:

* fast/block/float/float-with-anonymous-previous-sibling-expected.html: Added.
* fast/block/float/float-with-anonymous-previous-sibling.html: Added.

Identifier: 170874@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@194645 268f45cc-cd09-0410-ab3c-d52691b4dbfc