Directly-composited animated GIFs never resume once scrolled offscreen https://bugs.webkit.org/show_bug.cgi?id=152817 <rdar://problem/19982020>
Reviewed by Daniel Bates.

Source/WebCore:

Directly-composited animated GIFs would never resume once scrolled
offscreen. This is because calling repaint() in this case would not
cause BitmapImage::draw() to be called and the animation would thus
not be resumed. To address the problem,
repaintForPausedImageAnimationsIfNeeded() now calls
RenderBoxModelObject::contentChanged(ImageChanged) in addition to
repaint() to make sure the animation actually gets resumed, even in
the directly-composited animated GIF case.

Test: fast/images/composited-animated-gif-outside-viewport.html

* platform/graphics/BitmapImage.h:
Make currentFrame() public so it can be exposed via Internals for the
purpose of testing.

* rendering/RenderElement.cpp:
(WebCore::RenderElement::repaintForPausedImageAnimationsIfNeeded):
Call RenderBoxModelObject::contentChanged(ImageChanged) in addition to
calling repaint() to make sure the animation actually gets resumed in
the directly-composited animated GIFs case.

* testing/Internals.cpp:
(WebCore::Internals::imageFrameIndex):
* testing/Internals.h:
* testing/Internals.idl:
Expose new "unsigned long imageFrameIndex(Element)" operation on
Internals so layout tests can better check if an image is actually
animating. Previously, we would rely on the output of
internals.hasPausedImageAnimations(Element) but this is not sufficient
to cover this bug as our rendering code believed it has resumed the
animations but the GIF was not actually animating due to it being
directly-composited.

LayoutTests:

Add a layout test to check that directly-composited animated GIFs are
properly suspended / resumed based on visibility inside the viewport.

* fast/images/composited-animated-gif-outside-viewport-expected.txt: Added.
* fast/images/composited-animated-gif-outside-viewport.html: Added.


Identifier: 170926@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@194706 268f45cc-cd09-0410-ab3c-d52691b4dbfc